Improves hand selection.
Previously it was assumed that the 2 hole cards always had to be
present in the hand played by a player. Player::make_five has been
improved to consider hands where the 2 hole cards owned by a player
are not in the best hand that can be played by a player.

Improves TestPlayer::test_make_five.

Adds test fixture test_player.community_player for a test player that
has it's best hand in the community cards.

Adds full_house_community_cards.txt for sample input where a player
will have it's best hand in the community cards.